Serial # should be a long.  And Vector of longs not ints.

+++ src/cs5414/bank/message/BankRequestMessage.java
@@ -20,7 +20,7 @@ public class BankRequestMessage extends BaseMessage {
 	private static final long serialVersionUID = 1L;
 	
 	public RequestType requestType;
-	public int serial;
+	public long serial;
 	public String account;
 	public String accountInto;
 	public int amount;
